Move renderFooter & renderHeader to more appropriate classes
